# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateProjectRequest:
    properties:
      description:
        example: Redesign company website with new UI
        maxLength: 250
        type: string
      name:
        example: Website Redesign
        maxLength: 80
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dtos.UpdateProjectRequest:
    properties:
      description:
        example: Updated project scope
        maxLength: 250
        type: string
      name:
        example: Website Redesign v2
        maxLength: 80
        type: string
    type: object
...
  errors.BasicError:
    properties:
      error:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        maxLength: 250
        minLength: 5
        type: string
      id:
        type: string
      name:
        type: string
      org_id:
        type: string
      team_id:
        type: string
    required:
    - created_by
    - name
    - org_id
    - team_id
    type: object
  models.ProjectResult:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@taskio.com
    name: API Support
  description: Task management system API with real-time features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taskio API
  version: "1.0"
paths:
...
  /teams/{teamId}/projects:
    get:
      consumes:
      - application/json
      description: Retrieves all projects for a specific team
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Projects fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProjectResult'
                  type: array
              type: object
        "400":
          description: Invalid team ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a team member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get team projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Creates a new project within a team (owner or admin only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: Project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Projects
  /teams/{teamId}/projects/{projectId}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a project and all related tasks (owner only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Only owner can delete
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific project
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a team member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates project name and description
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      - description: Updated project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a team member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
securityDefinitions:
  BearerAuth:
    description: JWT token stored in cookie
    in: cookie
    name: taskio-cookie
    type: apiKey
swagger: "2.0"